SeafoodVision/ai/inference/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
onnxruntime==1.17.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==10.2.0
pydantic==2.6.4
python-multipart==0.0.9
structlog==24.1.0
